float error fix

+++ model_training.py
@@ -31,7 +31,7 @@ def train_base_model(model, dataloader, epochs=10):
         epoch_start_time = time.time()
         running_loss = 0.0
         for inputs, targets in tqdm(dataloader, desc=f"Epoch {epoch + 1}/{epochs}", disable=TQDM_DISABLE):
-            inputs, targets = inputs.to(device), targets.to(device)
+            inputs, targets = inputs.to(device), targets.to(device).float()
             optimizer.zero_grad()
             outputs = model(inputs)
             loss = criterion(outputs, targets)
